+++ tests/unit/test_dataset_geo.py
@@ -480,7 +480,11 @@ class TestDatasetGeo(TestCase):
             },
         ]
         expected = updates
+<<<<<<< add-source-dataset
         actual = [ast.literal_eval(a.decode("utf8")) for a in (snoop["payload"])]
+=======
+        actual = [json.loads(item) for item in snoop["payload"]]
+>>>>>>> master
         self.assertEqual(expected, actual)
 
         class NotAFeatureCollection:
@@ -491,7 +495,11 @@ class TestDatasetGeo(TestCase):
         snoop["payload"] = None
         nafc = NotAFeatureCollection()
         dataset.from_geo_features(nafc)
+<<<<<<< add-source-dataset
         actual = [ast.literal_eval(a.decode("utf8")) for a in (snoop["payload"])]
+=======
+        actual = [json.loads(item) for item in snoop["payload"]]
+>>>>>>> master
         self.assertEqual(expected, actual)
 
     @responses.activate
@@ -539,7 +547,11 @@ class TestDatasetGeo(TestCase):
             },
         ]
         expected = updates
+<<<<<<< add-source-dataset
         actual = [ast.literal_eval(a.decode("utf8")) for a in (snoop["payload"])]
+=======
+        actual = [json.loads(item) for item in snoop["payload"]]
+>>>>>>> master
         self.assertEqual(expected, actual)
 
     _dataset_json = {