+++ CHANGELOG.md
@@ -5,7 +5,8 @@
   - [#98](https://github.com/Datatamer/unify-client-python/issues/98) Add `__geo_interface__` to Dataset
   - [#100](https://github.com/Datatamer/unify-client-python/issues/100) Add `from_geo_features` to Dataset
   **BUG FIXES**
-  - [#62](https://github.com/Datatamer/unify-client-python/issues/62) `pairs`, `high_impact_pairs`, and `record_clusters` now fetch dataset information from the server
+   - [#118](https://github.com/Datatamer/unify-client-python/issues/118) Fix JSON sent for Dataset.update_records
+
 ## 0.4.0
   **BREAKING CHANGES**
   - [#61](https://github.com/Datatamer/unify-client-python/issues/61) `data` field renamed to `_data` (private).